+++ NukeValuable.cs
@@ -1,115 +1,142 @@
 using Photon.Pun;
 using UnityEngine;
 
-// TODO: Fix emissive warning texture overlay implementation when I have the time
-
+#pragma warning disable CS8618
 namespace LegalizeNuclearBombs
 {
     public class NukeValuable : Trap
     {
-        #pragma warning disable CS8618
-        
         public Transform center;
-
-        // public GameObject mesh;
-
-        // private Material _material;
-
-        // private bool _emissionActive;
-
-        // private float _emissionTime;
         
-        private ParticleScriptExplosion _particleScriptExplosion;
+        public GameObject mesh;
         
-        private int _hitCount;
+        private Material material;
+        private static readonly int emissionColor = Shader.PropertyToID("_EmissionColor");
+        
+        private bool detonated;
+        
+        private bool emissionActive;
+        private bool emissionImpulse;
+        
+        private float emissionTime;
+        
+        private ParticleScriptExplosion particleScriptExplosion;
+        
+        private int hitCount;
         
         public Sound warningSound;
         
         public override void Start()
         {
             base.Start();
-            _particleScriptExplosion = GetComponent<ParticleScriptExplosion>();
-            // _material = mesh.GetComponent<Renderer>().material;
+            particleScriptExplosion = GetComponent<ParticleScriptExplosion>();
+            material = mesh.GetComponent<Renderer>().material;
             
-            LegalizeNuclearBombs.Debug($"New nuke valuable spawned ({GetInstanceID()})");
+            LegalizeNuclearBombs.Debug($"New nuke valuable spawned | {gameObject}", this);
         }
-
-        // public void FixedUpdate()
-        // {
-        //     if (!_emissionActive) return;
-        //     _material.EnableKeyword("_EMISSION");
-        //     _emissionTime += Time.fixedDeltaTime;
-        //     
-        //     if (!(_emissionTime >= 1)) return;
-        //     _material.DisableKeyword("_EMISSION");
-        //     _emissionActive = false;
-        // }
+        
+        public override void Update()
+        {
+            // Debug
+            // if (Input.GetKeyDown(KeyCode.N))
+            // {
+            //     PlayWarningRPC();
+            // }
+            
+            if (!emissionActive) return;
+            if (emissionImpulse)
+            {
+                GameDirector.instance.CameraImpact.ShakeDistance(LegalizeNuclearBombs.configWarningCameraShakeStrength.Value, 1f, 6f, transform.position, .25f);
+                emissionImpulse = false;
+            }
+            LegalizeNuclearBombs.Debug($"emissionColor: {material.GetColor(emissionColor).r}");
+            material.SetColor(emissionColor, Color.white * Mathf.Lerp(
+                material.GetColor(emissionColor).r,
+                Mathf.Clamp(emissionTime, 0f, 1f),
+                .35f)
+            );
+            if (emissionTime > 0f)
+            {
+                emissionTime -= Time.deltaTime;
+                return;
+            }
+            if (material.GetColor(emissionColor).r > .01f) return;
+            material.SetColor(emissionColor, Color.black);
+            emissionActive = false;
+        }
         
         public void Explode()
         {
-            _particleScriptExplosion.Spawn(
+            if (detonated) return;
+            particleScriptExplosion.Spawn(
                 center.position,
-                LegalizeNuclearBombs.ConfigExplosionStrength.Value,
-                LegalizeNuclearBombs.ConfigPlayerDamage.Value,
-                LegalizeNuclearBombs.ConfigEnemyDamage.Value,
-                LegalizeNuclearBombs.ConfigExplosionStrength.Value,
+                LegalizeNuclearBombs.configExplosionStrength.Value,
+                LegalizeNuclearBombs.configPlayerDamage.Value,
+                LegalizeNuclearBombs.configEnemyDamage.Value,
+                LegalizeNuclearBombs.configExplosionStrength.Value,
                 false,
                 false,
-                LegalizeNuclearBombs.ConfigShakeMultiplier.Value
+                LegalizeNuclearBombs.configCameraShakeStrength.Value
                 );
-            LegalizeNuclearBombs.Debug($"KABOOM ({GetInstanceID()})");
+            LegalizeNuclearBombs.Debug("KABOOM", this);
+            detonated = true;
+            Destroy(gameObject); // Prevent spamming the explosion, especially at smaller strength
         }
         
-        public void PotentialExplodeLight() // Triggers if ConfigHitSensitivity is 2
+        #region UnityEvents
+        
+        public void PotentialExplodeLight()
         {
             if (!isLocal) return;
-            if (LegalizeNuclearBombs.ConfigHitSensitivity.Value > 1) PotentialExplodeHeavy();
+            if (LegalizeNuclearBombs.configHitSensitivity.Value is LegalizeNuclearBombs.HitSensitivity.Light)
+                PotentialExplodeHeavy();
         }
         
-        public void PotentialExplodeMedium() // Triggers if ConfigHitSensitivity is 1
+        public void PotentialExplodeMedium()
         {
             if (!isLocal) return;
-            if (LegalizeNuclearBombs.ConfigHitSensitivity.Value > 0) PotentialExplodeHeavy();
+            if (LegalizeNuclearBombs.configHitSensitivity.Value is not LegalizeNuclearBombs.HitSensitivity.Heavy)
+                PotentialExplodeHeavy();
         }
         
-        public void PotentialExplodeHeavy() // Always triggers
+        public void PotentialExplodeHeavy()
         {
-            if (!isLocal || LegalizeNuclearBombs.ConfigMaxHitCount.Value <= 0) return;
-            if (_hitCount >= LegalizeNuclearBombs.ConfigMaxHitCount.Value - 1)
+            if (!isLocal || LegalizeNuclearBombs.configMaxHitCount.Value <= 0) return;
+            if (hitCount >= LegalizeNuclearBombs.configMaxHitCount.Value - 1)
             {
                 Explode();
             }
             else
             {
-                LegalizeNuclearBombs.Debug($"_hitCount: {_hitCount + 1} ({GetInstanceID()})");
+                LegalizeNuclearBombs.Debug($"_hitCount: {hitCount + 1}", this);
                 // Play warning sound if almost about to go kaboom
-                if (_hitCount >= LegalizeNuclearBombs.ConfigMaxHitCount.Value - 2)
+                if (hitCount >= LegalizeNuclearBombs.configMaxHitCount.Value - 2)
                 {
-                    if (SemiFunc.IsMultiplayer())
-                        photonView.RPC(nameof(PlayWarningRPC), RpcTarget.All);
-                    else
-                        PlayWarningRPC();
+                    if (SemiFunc.IsMultiplayer()) photonView.RPC(nameof(PlayWarningRPC), RpcTarget.All);
+                    else PlayWarningRPC();
                 }
-                _hitCount++;
+                hitCount++;
             }
         }
         
-        // Configs checks are placed here so clients can have their own individual config preferences
+        #endregion
+        
         [PunRPC]
         public void PlayWarningRPC(PhotonMessageInfo info = default)
         {
-            if (LegalizeNuclearBombs.ConfigPlayWarningSound.Value)
+            if (LegalizeNuclearBombs.configPlayWarningSound.Value)
             {
-                warningSound.Volume = LegalizeNuclearBombs.ConfigWarningVolume.Value;
+                warningSound.Volume = LegalizeNuclearBombs.configWarningVolume.Value;
                 warningSound.Play(center.position);
-                LegalizeNuclearBombs.Debug($"Played audio warning; one hit left ({GetInstanceID()})");
             }
-
-            // if (LegalizeNuclearBombs.ConfigShowWarningTextureOverlay.Value)
-            // {
-            //     _emissionTime = 0f;
-            //     _emissionActive = true;
-            // }
+            
+            if (LegalizeNuclearBombs.configShowWarningVisual.Value)
+            {
+                emissionTime = 1.5f;
+                emissionImpulse = true;
+                emissionActive = true;
+            }
+            LegalizeNuclearBombs.Debug("Played warning, one hit left", this);
         }
     }
 }