+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Changelog
 
+### 2.0.0
+
+- Upgraded BepInEx dependency.
+- Implemented an emissive visual warning when the nuke has only one hit remaining.
+- Changed config entry `HitSensitivity` to use an enum value to make it more readable and intuitive, particularly with REPOConfig.
+- README changes:
+  - Updated the Configuration section to my new table style;
+  - Updated Discord URL to link to this mod's own dedicated thread.
+
 ### 1.0.0
+
 - Initial release 🎉